go button, timeout, turn scheduling, game deletion, user updates, refresh button
